Card Creation
Edited indenting on html.
Created a mock database in quoteApp.js.
Moved the Firebase initialization code to quoteApp.js.
Wrote a method called createAndDisplayCard that takes a quote object and creates html and fills in the data and displays the card to the view.
The createAndDisplayCard method is being called on startup using the data from the mock database.
Added a thumbs up image and a thumbs down image.